+++ src/main/java/forestry/apiculture/items/ItemHoneyComb.java
@@ -13,6 +13,8 @@ package forestry.apiculture.items;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 
+import net.minecraftforge.data.loading.DatagenModLoader;
+
 import forestry.api.core.ItemGroups;
 import forestry.core.items.ItemForestry;
 import forestry.core.items.definitions.IColoredItem;
@@ -21,7 +23,7 @@ public class ItemHoneyComb extends ItemForestry implements IColoredItem {
 	private final EnumHoneyComb type;
 
 	public ItemHoneyComb(EnumHoneyComb type) {
-		super(new Item.Properties().tab(type.isUnused() ? null : ItemGroups.tabApiculture));
+		super(new Item.Properties().tab((type.isUnused() && !DatagenModLoader.isRunningDataGen()) ? null : ItemGroups.tabApiculture));
 
 		this.type = type;
 	}