+++ changelog.md
@@ -4,6 +4,7 @@
 - Fixed loading of Chinese (simplified) language file (#44)
 - Fixed Apiarist chest menu not displaying queens/princesses/drones in menu (#40)
 - Unused honeycombs are no longer obtainable in survival, from village houses or trades (#42)
+- Fix fluid attributes and use honey/milk tags where possible
 
 ## Forestry 1.0.3
 - Fixed saplings not dropping themselves (#36)